Add websocket support for 2016 TVs
