+++ README.md
@@ -1,12 +1,14 @@
 samsungctl
 ==========
 samsungctl is a library and a command line tool for remote controlling Samsung
-televisions via a TCP/IP connection. It should work with any modern Samsung TV
-with Ethernet or Wi-Fi connectivity.
+televisions via a TCP/IP connection. It currently supports both pre-2016 TVs
+as well most of the modern Tizen-OS TVs with Ethernet or Wi-Fi connectivity.
 
 Dependencies
 ------------
-samsungctl requires Python 3, but no additional libraries.
+
+- `Python 3`
+- `websocket-client` (optional, for 2016+ TVs)
 
 Installation
 ------------
@@ -57,6 +59,7 @@ optional arguments:
   -i, --interactive   interactive control
   --host HOST         TV hostname or IP address
   --port PORT         TV port number (TCP)
+  --method METHOD     Connection method (legacy or websocket)
   --name NAME         remote control name
   --description DESC  remote control description
   --id ID             remote control id
@@ -90,14 +93,15 @@ with samsungctl.Remote(config) as remote:
 The constructor takes a configuration dictionary as a parameter. All
 configuration items must be specified.
 
-| Key         | Type   | Description                               |
-| ----------- | ------ | ----------------------------------------- |
-| host        | string | Hostname or IP address of the TV.         |
-| port        | int    | TCP port number. (Default: `55000`)       |
-| name        | string | Name of the remote controller.            |
-| description | string | Remote controller description.            |
-| id          | string | Additional remote controller ID.          |
-| timeout     | int    | Timeout in seconds. `0` means no timeout. |
+| Key         | Type   | Description                                 |
+| ----------- | ------ | -----------------------------------------   |
+| host        | string | Hostname or IP address of the TV.           |
+| port        | int    | TCP port number. (Default: `55000`)         |
+| method      | string | Connection method ("legacy" or "websocket") |
+| name        | string | Name of the remote controller.              |
+| description | string | Remote controller description.              |
+| id          | string | Additional remote controller ID.            |
+| timeout     | int    | Timeout in seconds. `0` means no timeout.   |
 
 The `Remote` object is very simple and you only need the `control(key)` method.
 The only parameter is a string naming the key to be sent (e.g.
@@ -130,6 +134,7 @@ config = {
     "id": "",
     "host": "192.168.0.10",
     "port": 55000,
+    "method": "legacy",
     "timeout": 0,
 }
 
@@ -183,6 +188,8 @@ list has some common key codes and their descriptions.
 | KEY_DTV           | TV Source    |
 | KEY_HDMI          | HDMI Source  |
 
+Please note that some codes are different on the 2016+ TVs. For example, `KEY_POWEROFF` is `KEY_POWER` on the newer TVs.
+
 References
 ----------
 I did not reverse engineer the control protocol myself and samsungctl is not
@@ -191,3 +198,4 @@ the only implementation. Here is the list of things that inspired samsungctl.
 - http://sc0ty.pl/2012/02/samsung-tv-network-remote-control-protocol/
 - https://gist.github.com/danielfaust/998441
 - https://github.com/Bntdumas/SamsungIPRemote
+- https://github.com/kyleaa/homebridge-samsungtv2016
