+++ requirements.txt
@@ -1,0 +1,1 @@
+websocket-client==0.37.0
