+++ samsungctl.conf
@@ -2,6 +2,7 @@
 	"name": "samsungctl",
 	"description": "PC",
 	"id": "",
+	"method": "legacy",
 	"port": 55000,
 	"timeout": 0
 }
